split bin into lib/bin
